--- som/plot/circle.py
@@ -1,34 +1,0 @@
-# -*- coding: utf-8 -*-
-
-#    Visigoth: A lightweight Python3 library for rendering data visualizations in SVG
-#    Copyright (C) 2020-2021  Visigoth Developers
-#
-#   Permission is hereby granted, free of charge, to any person obtaining a copy of this software
-#   and associated documentation files (the "Software"), to deal in the Software without
-#   restriction, including without limitation the rights to use, copy, modify, merge, publish,
-#   distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the
-#   Software is furnished to do so, subject to the following conditions:
-#
-#   The above copyright notice and this permission notice shall be included in all copies or
-#   substantial portions of the Software.
-#
-#   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING
-#   BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-#   NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
-#   DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-#   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
-from .svgstyled import SvgStyled
-
-# represent a circle as an SVG object
-class Circle(SvgStyled):
-
-    def __init__(self,x,y,r,fill="grey",tooltip="",stroke=None,stroke_width=None):
-        super().__init__('circle',tooltip)
-        self.addAttr("cx",x).addAttr("cy",y).addAttr("r",r)
-        if fill:
-            self.addAttr("fill",fill)
-        if stroke:
-            self.addAttr("stroke",stroke)
-        if stroke_width != None:
-            self.addAttr("stroke-width",stroke_width)
