--- som/plot/hexagon.py
@@ -1,40 +1,0 @@
-# MIT License
-#
-# Copyright (c) 2022 Niall McCarroll
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-from .polygon import Polygon
-
-import math
-
-class Hexagon(Polygon):
-
-    def __init__(self,x,y,dlength,fill,stroke,stroke_width):
-        points = []
-
-        r = 0
-        for idx in range(0,6):
-            py = y+dlength*math.cos(r)
-            px = x+dlength*math.sin(r)
-            points.append((px,py))
-            r += math.pi/3
-
-        self.hexpoints = points
-        super().__init__(points,fill,stroke,stroke_width)
